src/modes/execute/requirements.txt
i3ipc==2.2.1
libtmux==0.15.1
plug[qplug]==0.0.1
